Adds liftToHeight Code to 24 Inch Worlds
Implements lift to height code in 24 inch worlds with proportional only
controller based on the average of the two potentiometers to go to the
height of three posts.

Also adds the memory leak elimination project which was supposed to be
used on the test robot but testing proved it not to be useful at this
time.